Add colors from react palette library. Create CriptoCard component
